+++ frontend/src/components/PostsDisplay/PostsDisplay.module.css
@@ -1,0 +1,11 @@
+.postBodyArea {
+  background-color: var(--soft-background);
+  color: white;
+  font-size: 1rem;
+  line-height: 1.5;
+  padding: 15px;
+  margin: 15px auto;
+  width: 95%;
+  border: var(--light-border);
+  border-radius: 15px;
+}
