+++ frontend/src/styles/shared.module.css
@@ -10,15 +10,3 @@
   padding: 0;
   width: 0;
 }
-
-.postBodyArea {
-  background-color: var(--soft-background);
-  color: white;
-  font-size: 1rem;
-  line-height: 1.5;
-  padding: 15px;
-  margin: 15px;
-  width: 95%;
-  border: var(--light-border);
-  border-radius: 15px;
-}
